+++ permuter/core/__init__.py
