+++ permuter/core/algo.py
